requirements.txt
# ITENERO - AI-Powered Travel Itinerary Generator
# Requirements file for Python dependencies

# Core Framework
Flask==2.3.3
Werkzeug==2.3.7

# AI Integration
google-generativeai==0.3.2

# HTTP Requests
requests==2.31.0

# Email Support
Flask-Mail==0.9.1

# OAuth Integration
Flask-Dance==7.0.0

# Additional Dependencies
python-dotenv==1.0.0
gunicorn==21.2.0

# Development Dependencies (optional)
pytest==7.4.3
pytest-flask==1.3.0
black==23.9.1
flake8==6.1.0
